+++ packages/fe-fpm-writer/src/filter/index.ts
@@ -55,7 +55,7 @@ export async function generateCustomFilter(basePath: string, filterConfig: Custo
         fs = create(createStorage());
     }
     await validateBasePath(basePath, fs);
-    const fnGenerateId = await createIdGenerator(basePath, fs);
+    const fnGenerateId = await createIdGenerator({ basePath, fsEditor: fs });
 
     const { path: manifestPath, content: manifest } = await getManifest(basePath, fs);
     const config = enhanceConfig(filterConfig, manifestPath, manifest);