+++ packages/fiori-annotation-api/src/external-services.ts
@@ -2,6 +2,49 @@ import { join } from 'node:path';
 import { join as joinPosix } from 'node:path/posix';
 import type { ValueListReference } from './types/adapter';
 import { readFile, access, constants } from 'node:fs/promises';
+
+/**
+ * Adapts the given segment by removing the alias or namespace, or replacing alias with the namespace in action or function parameters.
+ *
+ * @param segment - The segment to adapt, which is a part of the metadata path for an external service.
+ * @param namespace - The namespace of the annotation file where the external service reference is defined.
+ * @param alias - The alias of the annotation file where the external service reference is defined, if any.
+ * @returns - The adapted segment with the alias or namespace removed, or alias replaced with the namespace in action or function parameters.
+ */
+function adaptSegment(segment: string, namespace: string, alias: string | undefined): string {
+    const [base, ...params] = segment.split('(');
+    let adaptedSegment = base.split('.').pop() ?? '';
+    if (params.length) {
+        const adaptedParams = params
+            .join('(')
+            .split(',')
+            .map((param) => {
+                let adaptedParam = param;
+                let isCollection = false;
+                if (param.includes('Collection(')) {
+                    isCollection = true;
+                    adaptedParam = param.replace('Collection(', '');
+                }
+                const padding = new RegExp(/^\s*/).exec(adaptedParam)?.[0] ?? '';
+                const paramText = adaptedParam.trim();
+                const segments = paramText.split('.');
+                adaptedParam = segments.pop()!;
+                if (segments.length) {
+                    let prefix = segments.join('.').trim();
+                    prefix = prefix === namespace || (alias && prefix === alias) ? namespace : prefix;
+                    adaptedParam = padding + prefix + '.' + adaptedParam;
+                }
+                if (isCollection) {
+                    adaptedParam = 'Collection(' + adaptedParam;
+                }
+                return adaptedParam;
+            })
+            .join(',');
+        adaptedSegment += '(' + adaptedParams;
+    }
+    return adaptedSegment;
+}
+
 /**
  * Reads the metadata of external services based on the provided definitions and returns a map of the service URI to its metadata content and local file path.
  *
@@ -26,7 +69,7 @@ export async function readExternalServiceMetadata(
                     relativeServicePath,
                     target
                         .split('/')
-                        .map((segment) => segment.split('.').pop())
+                        .map((segment) => adaptSegment(segment, reference.namespace, reference.alias))
                         .join('/')
                 );