+++ packages/cds-odata-annotation-converter/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/cds-odata-annotation-converter
 
+## 0.7.14
+
+### Patch Changes
+
+- 55417bb: fix(deps): update dependency i18next to v25.8.18
+
 ## 0.7.13
 
 ### Patch Changes