+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @sap-ux/fe-fpm-writer
 
+## 0.43.14
+
+### Patch Changes
+
+- Updated dependencies [f384ace]
+    - @sap-ux/project-access@1.35.15
+    - @sap-ux/fiori-annotation-api@0.9.36
+
+## 0.43.13
+
+### Patch Changes
+
+- 08f3a5c: Refactored createIdGenerator to use object parameters for reusing for v2 scenarios.
+
+## 0.43.12
+
+### Patch Changes
+
+- @sap-ux/project-access@1.35.14
+- @sap-ux/fiori-annotation-api@0.9.35
+
+## 0.43.11
+
+### Patch Changes
+
+- Updated dependencies [436cad8]
+    - @sap-ux/fiori-annotation-api@0.9.34
+
+## 0.43.10
+
+### Patch Changes
+
+- 55417bb: fix(deps): update dependency i18next to v25.8.18
+    - @sap-ux/fiori-annotation-api@0.9.33
+
 ## 0.43.9
 
 ### Patch Changes