+++ packages/fiori-freestyle-writer/CHANGELOG.md
@@ -1,5 +1,63 @@
 # @sap-ux/fiori-freestyle-writer
 
+## 2.5.79
+
+### Patch Changes
+
+- @sap-ux/cap-config-writer@0.12.81
+- @sap-ux/fiori-generator-shared@0.13.89
+- @sap-ux/odata-service-writer@0.31.1
+- @sap-ux/ui5-application-writer@1.8.1
+- @sap-ux/ui5-test-writer@0.7.92
+
+## 2.5.78
+
+### Patch Changes
+
+- Updated dependencies [25e5177]
+    - @sap-ux/odata-service-writer@0.31.0
+    - @sap-ux/ui5-config@0.30.0
+    - @sap-ux/ui5-application-writer@1.8.1
+    - @sap-ux/cap-config-writer@0.12.80
+    - @sap-ux/fiori-generator-shared@0.13.88
+    - @sap-ux/ui5-test-writer@0.7.91
+
+## 2.5.77
+
+### Patch Changes
+
+- Updated dependencies [b516f01]
+    - @sap-ux/ui5-test-writer@0.7.90
+
+## 2.5.76
+
+### Patch Changes
+
+- Updated dependencies [ae6758f]
+    - @sap-ux/fiori-generator-shared@0.13.87
+    - @sap-ux/cap-config-writer@0.12.79
+
+## 2.5.75
+
+### Patch Changes
+
+- Updated dependencies [b5c7d47]
+    - @sap-ux/ui5-application-writer@1.8.0
+    - @sap-ux/ui5-test-writer@0.7.89
+
+## 2.5.74
+
+### Patch Changes
+
+- 55417bb: fix(deps): update dependency i18next to v25.8.18
+- Updated dependencies [55417bb]
+    - @sap-ux/cap-config-writer@0.12.78
+    - @sap-ux/fiori-generator-shared@0.13.86
+    - @sap-ux/odata-service-writer@0.30.1
+    - @sap-ux/ui5-application-writer@1.7.20
+    - @sap-ux/ui5-test-writer@0.7.88
+    - @sap-ux/ui5-config@0.29.21
+
 ## 2.5.73
 
 ### Patch Changes