+++ packages/mockserver-config-writer/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @sap-ux/mockserver-config-writer
 
+## 0.9.65
+
+### Patch Changes
+
+- Updated dependencies [f384ace]
+    - @sap-ux/project-access@1.35.15
+
+## 0.9.64
+
+### Patch Changes
+
+- Updated dependencies [25e5177]
+    - @sap-ux/ui5-config@0.30.0
+    - @sap-ux/project-access@1.35.14
+
+## 0.9.63
+
+### Patch Changes
+
+- 55417bb: fix(deps): update dependency i18next to v25.8.18
+    - @sap-ux/ui5-config@0.29.21
+
 ## 0.9.62
 
 ### Patch Changes