+++ packages/ui-prompting/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/ui-prompting
 
+## 0.6.16
+
+### Patch Changes
+
+- Updated dependencies [7f2689f]
+    - @sap-ux/ui-components@2.1.3
+
 ## 0.6.15
 
 ### Patch Changes