+++ packages/ui5-library-reference-writer/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/ui5-library-reference-writer
 
+## 0.2.62
+
+### Patch Changes
+
+- Updated dependencies [f384ace]
+    - @sap-ux/project-access@1.35.15
+
+## 0.2.61
+
+### Patch Changes
+
+- Updated dependencies [25e5177]
+    - @sap-ux/ui5-config@0.30.0
+    - @sap-ux/project-access@1.35.14
+
 ## 0.2.60
 
 ### Patch Changes